functions/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-identity
azure-communication-messages
openai
python-dotenv
aiohttp
azure-ai-projects>=1.0.0b12